+++ api/microchip/pic16f87x/pin_fn/pin_fn.c.v
@@ -5,30 +5,30 @@
 // Pin management functions for 16F87x
 module pin_fn
 
-@[as_macro] pub const a0 = 0 // PORTA pins
-@[as_macro] pub const a1 = 1
-@[as_macro] pub const a2 = 2
-@[as_macro] pub const a3 = 3
-@[as_macro] pub const a4 = 4
-@[as_macro] pub const a5 = 5
+@[as_macro] pub const a0 = u8(0) // PORTA pins
+@[as_macro] pub const a1 = u8(1)
+@[as_macro] pub const a2 = u8(2)
+@[as_macro] pub const a3 = u8(3)
+@[as_macro] pub const a4 = u8(4)
+@[as_macro] pub const a5 = u8(5)
 
-@[as_macro] pub const b0 = 8 // PORTB pins
-@[as_macro] pub const b1 = 9
-@[as_macro] pub const b2 = 10
-@[as_macro] pub const b3 = 11
-@[as_macro] pub const b4 = 12
-@[as_macro] pub const b5 = 13
-@[as_macro] pub const b6 = 14
-@[as_macro] pub const b7 = 15
+@[as_macro] pub const b0 = u8(8) // PORTB pins
+@[as_macro] pub const b1 = u8(9)
+@[as_macro] pub const b2 = u8(10)
+@[as_macro] pub const b3 = u8(11)
+@[as_macro] pub const b4 = u8(12)
+@[as_macro] pub const b5 = u8(13)
+@[as_macro] pub const b6 = u8(14)
+@[as_macro] pub const b7 = u8(15)
 
-@[as_macro] pub const c0 = 16 // PORTC pins
-@[as_macro] pub const c1 = 17
-@[as_macro] pub const c2 = 18
-@[as_macro] pub const c3 = 19
-@[as_macro] pub const c4 = 20
-@[as_macro] pub const c5 = 21
-@[as_macro] pub const c6 = 22
-@[as_macro] pub const c7 = 23
+@[as_macro] pub const c0 = u8(16) // PORTC pins
+@[as_macro] pub const c1 = u8(17)
+@[as_macro] pub const c2 = u8(18)
+@[as_macro] pub const c3 = u8(19)
+@[as_macro] pub const c4 = u8(20)
+@[as_macro] pub const c5 = u8(21)
+@[as_macro] pub const c6 = u8(22)
+@[as_macro] pub const c7 = u8(23)
 
 @[as_macro] pub const d0 = u8(24) // PORTA pins
 @[as_macro] pub const d1 = u8(25)
@@ -39,9 +39,9 @@ module pin_fn
 @[as_macro] pub const d6 = u8(30)
 @[as_macro] pub const d7 = u8(31)
 
-@[as_macro] pub const e0 = 32
-@[as_macro] pub const e1 = 33
-@[as_macro] pub const e2 = 34
+@[as_macro] pub const e0 = u8(32)
+@[as_macro] pub const e1 = u8(33)
+@[as_macro] pub const e2 = u8(34)
 
 // IRQ (INT) pin
 @[as_macro] pub const irq = u8(8) // RB0/INT
